# api/swagger/swagger.yaml
definitions:
  controllers.LoginByUsernamePasswordRequest:
    properties:
      password:
        default: "123456"
        type: string
      username:
        default: admin
        type: string
    type: object
info:
  contact:
    name: Dorvan
    url: https://github.com/Xuduoteng
  description: This is a swagger api document for Go Gin Boilerplate.
  title: Go Gin Boilerplate API
  version: v1.0.0
paths:
  /examples/createExample:
    post:
      description: Create Example
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Example
  /examples/deleteExample:
    post:
      description: Delete Example
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Example
  /examples/getExample:
    get:
      description: Get Example
      parameters:
      - description: the example id
        in: query
        name: exampleId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Example
  /examples/updateExample:
    post:
      description: Update Example
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Example
  /public/ping:
    get:
      description: Ping
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - Public
  /users/loginByUsernamePassword:
    post:
      description: Login By Username Password
      parameters:
      - description: username、password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginByUsernamePasswordRequest'
      responses: {}
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"